* convert to int;

+++ include/BaseControl.hpp
@@ -11,7 +11,7 @@ public:
 	explicit BaseControl(PlayGround& playGround, unsigned short startTileId);
 	BaseControl(const BaseControl&) = delete;
 	BaseControl& operator= (const BaseControl&) = delete;
-	const Vector2D<float>& GetPos() const;
+	const Vector2D<int>& GetPos() const;
 	virtual void Update(float deltaTime, IDrawable* pDrawable);
 	virtual ~BaseControl() = default;
 
@@ -21,7 +21,7 @@ protected:
 	void Move(const Vector2D<signed short>& dir, float deltaTime);
 
 private:
-	Vector2D<float> m_pos{0.f, 0.f};
+	Vector2D<int> m_pos{0, 0};
 	unsigned short m_currentTileId{0};
 };
